requirements.txt
streamlit
requests
google-generativeai
unidecode
fpdf2